Split experimentation code out from unit tests
