+++ src/platformio.ini
@@ -4,6 +4,8 @@
 extra_configs =
 	# keep common first
 	targets/common.ini
+	# experimentation targets
+	targets/experiments.ini
 	# defined one by one to maintain the order
 	targets/frsky.ini
 	targets/happymodel_900.ini
