+++ src/src/experiments/crc_nonce_recovery.cpp
@@ -1,5 +1,4 @@
 #include <cstdint>
-#include <unity.h>
 #include <crc.h>
 #include "common.h"
