+++ src/targets/common.ini
@@ -12,7 +12,7 @@ monitor_dtr = 0
 monitor_rts = 0
 
 [common_env_data]
-src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<test/> -<tests/> -<*.py> -<*test*.*>
+src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<test/> -<tests/> -<*.py> -<*test*.*> -<experiments/>
 build_flags = -Wall -Iinclude
 build_flags_tx = -DTARGET_TX=1 ${common_env_data.build_flags}
 build_flags_rx = -DTARGET_RX=1 ${common_env_data.build_flags}
