+++ src/targets/experiments.ini
@@ -1,0 +1,31 @@
+#
+# These targets produce native applications that are used to generate data for verifying assumptions and trying experiments
+#
+
+[env:crc_nonce_recovery]
+platform = native
+framework =
+lib_ignore = BUTTON, DAC, EEPROM, LQCALC, POWERMGNT
+src_filter = +<experiments/crc_nonce_recovery.cpp> -<ESP32*.*> -<STM32*.*> -<ESP8*.*> -<tx_*.cpp> -<rx_*.cpp> -<common.*> -<config.*>
+build_flags =
+	-std=c++11
+	-Iinclude
+	-D UNIT_TEST
+	-D TARGET_NATIVE
+	-D ENABLE_TELEMETRY
+	-D CRSF_RX_MODULE
+	-D CRSF_TX_MODULE
+
+[env:crc_experiments]
+platform = native
+framework =
+lib_ignore = BUTTON, DAC, EEPROM, LQCALC, POWERMGNT
+src_filter = +<experiments/crc_experiments.cpp> -<ESP32*.*> -<STM32*.*> -<ESP8*.*> -<tx_*.cpp> -<rx_*.cpp> -<common.*> -<config.*>
+build_flags =
+	-std=c++11
+	-Iinclude
+	-D UNIT_TEST
+	-D TARGET_NATIVE
+	-D ENABLE_TELEMETRY
+	-D CRSF_RX_MODULE
+	-D CRSF_TX_MODULE
